Overhaul

+++ engine/src/render_resource.cpp
@@ -5,7 +5,7 @@ using namespace MilSim;
 /**
  * TEXTURE
  */
-void RenderResourceContext::push_texture(TextureData tex, size_t ref)
+void RenderResourceContext::push_texture(TextureData tex, uint64_t ref)
 {
 	m_tex.push_back(tex);
 	m_tex_ref.push_back(ref);
@@ -14,7 +14,7 @@ void RenderResourceContext::push_texture(TextureData tex, size_t ref)
 /**
  * VERTEX BUFFER
  */
-void RenderResourceContext::push_vertex_buffer(VertexBufferData vb, size_t ref)
+void RenderResourceContext::push_vertex_buffer(VertexBufferData vb, uint64_t ref)
 {
 	m_vb.push_back(vb);
 	m_vb_ref.push_back(ref);
@@ -23,7 +23,7 @@ void RenderResourceContext::push_vertex_buffer(VertexBufferData vb, size_t ref)
 /**
  * VERTEX LAYOUT
  */
-void RenderResourceContext::push_vertex_layout(VertexLayoutData vl, size_t ref)
+void RenderResourceContext::push_vertex_layout(VertexLayoutData vl, uint64_t ref)
 {
 	m_vl.push_back(vl);
 	m_vl_ref.push_back(ref);
@@ -32,7 +32,7 @@ void RenderResourceContext::push_vertex_layout(VertexLayoutData vl, size_t ref)
 /**
  * INDEX BUFFER
  */
-void RenderResourceContext::push_index_buffer(RenderResourceContext::IndexBufferData ib, size_t ref)
+void RenderResourceContext::push_index_buffer(RenderResourceContext::IndexBufferData ib, uint64_t ref)
 {
 	m_ib.push_back(ib);
 	m_ib_ref.push_back(ref);
